// examples/templates-advanced/generated/ffmpeg-rest-api.ts
export const FFMPEG_REST_API_DEFAULT_CONFIG = {
  services: {
    '0b194aa4-20ba-4e4a-add6-90c0bd821f07': {
      icon: 'https://cdn.sanity.io/images/sy1jschh/production/0ce0bfdcfbdbf69662b1116671f97c2dd788b655-157x157.svg',
      name: 'Redis',
      deploy: {
        startCommand:
          '/bin/sh -c "rm -rf $RAILWAY_VOLUME_MOUNT_PATH/lost+found/ && exec docker-entrypoint.sh redis-server --requirepass $REDIS_PASSWORD --save 60 1 --dir $RAILWAY_VOLUME_MOUNT_PATH"',
      },
      source: {
        image: 'redis:7-alpine',
      },
      variables: {
        REDISHOST: {
          isOptional: false,
          defaultValue: '${{RAILWAY_PRIVATE_DOMAIN}}',
        },
        REDISPORT: {
          isOptional: false,
          defaultValue: '6379',
        },
        REDISUSER: {
          isOptional: false,
          defaultValue: 'default',
        },
        REDIS_URL: {
          isOptional: false,
          description: 'Connection string for connecting to redis using the private network',
          defaultValue:
            'redis://${{ REDISUSER }}:${{ REDIS_PASSWORD }}@${{ REDISHOST }}:${{ REDISPORT }}',
        },
        REDISPASSWORD: {
          isOptional: false,
          defaultValue: '${{REDIS_PASSWORD}}',
        },
        REDIS_PASSWORD: {
          isOptional: false,
          defaultValue: '${{ secret(32, "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ") }}',
        },
        REDIS_PUBLIC_URL: {
          description: 'Connection string for connecting to redis externally',
          defaultValue:
            'redis://default:${{ REDIS_PASSWORD }}@${{ RAILWAY_TCP_PROXY_DOMAIN }}:${{ RAILWAY_TCP_PROXY_PORT }}',
        },
      },
      networking: {
        tcpProxies: {
          '6379': {},
        },
      },
      volumeMounts: {
        '0b194aa4-20ba-4e4a-add6-90c0bd821f07': {
          mountPath: '/data',
        },
      },
    },
    // Service ID from the generated config - use this to access service-specific variables with full type safety
    '7a5f33d0-fc76-4ba3-a311-b46b17f050f0': {
      name: 'ffmpeg-rest',
      source: {
        repo: 'crisog/ffmpeg-rest',
        branch: null,
        rootDirectory: null,
      },
      variables: {
        PORT: {
          defaultValue: '3000',
        },
        NODE_ENV: {
          defaultValue: 'production',
        },
        TEMP_DIR: {
          defaultValue: '/tmp/ffmpeg-rest',
        },
        REDIS_URL: {
          description: '',
          defaultValue: '${{Redis.REDIS_URL}}',
        },
        S3_BUCKET: {
          isOptional: true,
          description: '',
          defaultValue: '',
        },
        S3_REGION: {
          isOptional: true,
          description: '',
          defaultValue: '',
        },
        S3_ENDPOINT: {
          isOptional: true,
          description: '',
          defaultValue: '',
        },
        STORAGE_MODE: {
          description:
            "Set to `s3` if you want to enable S3 mode. You'll need to configure S3 variables.",
          defaultValue: '',
        },
        MAX_FILE_SIZE: {
          defaultValue: '104857600',
        },
        S3_PUBLIC_URL: {
          isOptional: true,
          description: 'Publicly accessible URL to the S3 bucket',
          defaultValue: '',
        },
        S3_PATH_PREFIX: {
          isOptional: true,
          description: 'Subdirectory where files will be stored within the bucket',
          defaultValue: '',
        },
        S3_ACCESS_KEY_ID: {
          isOptional: true,
          description: '',
          defaultValue: '',
        },
        S3_DEDUP_ENABLED: {
          isOptional: true,
          description: '',
          defaultValue: 'true',
        },
        S3_DEDUP_TTL_DAYS: {
          isOptional: true,
          description: '',
          defaultValue: '90',
        },
        S3_SECRET_ACCESS_KEY: {
          isOptional: true,
          description: '',
          defaultValue: '',
        },
      },
    },
  },
};